Refactor del archivo de rutas: se separó en tres archivos separados. Agregada la función de eliminación de cuida. Solucionado el error al cargar la lista de ciudades
