+++ flapper-news/public/javascripts/angularApp.js
@@ -53,21 +53,21 @@ app.factory('auth', ['$http', '$window', function($http, $window){
 
 app.factory('trips', ['$http', 'auth', function($http, auth){
 	var o = {
-		trips: []
+		trips: [],
+		cities: []
 	};
 
 	o.getAll = function() {
 		return $http.get('/trips', {
-				headers: {Authorization: 'Bearer '+auth.getToken()}
+				headers: {Authorization: 'Bearer '+ auth.getToken()}
 			}).success(function(data){
 			angular.copy(data, o.trips);
 		});
 	};
 
-
 	o.create = function(trip) {
 		return $http.post('/trips', trip, {
-			headers: {Authorization: 'Bearer '+auth.getToken()}
+			headers: {Authorization: 'Bearer '+ auth.getToken()}
 		}).success(function(data){
 			o.trips.push(data);
 		});
@@ -75,8 +75,6 @@ app.factory('trips', ['$http', 'auth', function($http, auth){
 
 	o.get = function(id) {
 		return $http.get('/trips/' + id).then(function(res){
-			console.log("BBBBBBBBH");
-			console.log(res.data);
 			return res.data;
 		});
 	};
@@ -87,12 +85,50 @@ app.factory('trips', ['$http', 'auth', function($http, auth){
 
 	o.addCity = function(id, city) {
 		return $http.post('/trips/' + id + '/cities', city, {
-		    headers: {Authorization: 'Bearer '+auth.getToken()}
-		  });
+		    headers: {Authorization: 'Bearer '+ auth.getToken()}
+		  }).success(function(data){
+			o.cities.push(data);
+		});
 	  };
+
+	o.removeCity = function(id, city) {
+		return $http.post('/trips/' + id + '/cities/' + city + '/remove', {
+			headers: {Authorization: 'Bearer '+ auth.getToken()}}
+		);
+	};
 
 	return o;
 }]);
+
+app.controller('AuthCtrl', [ '$scope', '$state', 'auth',
+	function($scope, $state, auth){
+		$scope.user = {};
+
+		$scope.register = function(){
+			auth.register($scope.user).error(function(error){
+				$scope.error = error;
+			}).then(function(){
+				$state.go('home');
+			});
+		};
+
+		$scope.logIn = function(){
+			auth.logIn($scope.user).error(function(error){
+				$scope.error = error;
+			}).then(function(){
+				$state.go('home');
+			});
+		};
+	}
+]);
+
+app.controller('NavCtrl', [ '$scope', 'auth',
+	function($scope, auth){
+		$scope.isLoggedIn = auth.isLoggedIn;
+		$scope.currentUser = auth.currentUser;
+		$scope.logOut = auth.logOut;
+	}
+]);
 
 app.controller('MainCtrl', [ '$scope', '$window', 'trips', 'auth',
 	function($scope, $window, trips, auth){
@@ -106,7 +142,7 @@ app.controller('MainCtrl', [ '$scope', '$window', 'trips', 'auth',
 				description: $scope.description,
 				dateOfDeparture: $scope.dateOfDeparture,
 				arrivalDate: $scope.arrivalDate,
-				author: auth.currentUser,
+				author: auth.currentUser
 			});
 			$scope.name = '';
 			$scope.description = '';
@@ -128,66 +164,54 @@ app.controller('MainCtrl', [ '$scope', '$window', 'trips', 'auth',
 	}
 ]);
 
-app.controller('TripsCtrl', [ '$scope', 'trips', 'trip', 'auth', '$state',
-	function($scope, trips, trip, auth, $state){
+app.controller('TripsCtrl', [ '$scope',  '$window', 'trips', 'trip', 'auth',
+	function($scope,  $window, trips, trip, auth){
 		$scope.trip = trip;
+		$scope.cities = trip.cities;
 		$scope.isLoggedIn = auth.isLoggedIn;
 
 		// Agrego esta logica para centrar el mapa siempre que hayan cuidades guardadas.
 		// Al tener varias cuidades necesitamos tener una de referencia para centrar el mapa.
-		if(!$scope.trip.cities || $scope.trip.cities=== '' || $scope.trip.cities.length<1) {
-			$scope.map = { center: { latitude: -34.6037232, longitude:  -58.38159310000003}, zoom: 8 }; 
-		}else{
-			var city = $scope.trip.cities.pop();
-			$scope.map = { center: { latitude: city.latitude, longitude:  city.longitude}, zoom: 8 }; 
+
+		if(!$scope.cities || $scope.cities === '' || $scope.cities.length < 1) {
+			$scope.map = { center: { latitude: -34.6037232, longitude:  -58.38159310000003}, zoom: 8 };
+
+		} else {
+			centerMap();
 		}
 
-
 		$scope.addCity = function(){
-		  if($scope.city === '') { return;}
+			if($scope.city === '') { return; }
 
-		  trips.addCity(trip._id, {
-		    address: $scope.city.formatted_address,
-		    longitude: $scope.city.geometry.location.F, 
-		    latitude: $scope.city.geometry.location.A,
-		    icon: $scope.city.icon,
-		  }).success(function(city) {
-		    $scope.trip.cities.push(city);
-		  });
-
-		  $scope.city = '';
-		};
-
-	}
-]);
-
-app.controller('AuthCtrl', [ '$scope', '$state', 'auth',
-	function($scope, $state, auth){
-		$scope.user = {};
-
-		$scope.register = function(){
-			auth.register($scope.user).error(function(error){
-			  $scope.error = error;
-			}).then(function(){
-			  $state.go('home');
+			trips.addCity(trip._id, {
+				address: $scope.city.formatted_address,
+				longitude: $scope.city.geometry.location.F,
+				latitude: $scope.city.geometry.location.A,
+				icon: $scope.city.icon
+			}).success(function(city) {
+				$scope.cities.push(city);
+				centerMap();
 			});
+			$scope.city = '';
 		};
 
-		$scope.logIn = function(){
-			auth.logIn($scope.user).error(function(error){
-			  $scope.error = error;
-			}).then(function(){
-			  $state.go('home');
-			});
+		$scope.removeCity = function(city){
+			var deleteCity = $window.confirm('Are you sure you want to delete?');
+			if(deleteCity){
+				trips.removeCity(trip._id, city._id);
+				var _city = $scope.cities.indexOf(city);
+				$scope.cities.splice(_city, 1);
+			}
 		};
-	}
-]);
 
-app.controller('NavCtrl', [ '$scope', 'auth',
-	function($scope, auth){
-		$scope.isLoggedIn = auth.isLoggedIn;
-		$scope.currentUser = auth.currentUser;
-		$scope.logOut = auth.logOut;
+		$scope.viewMap = false;
+
+		function centerMap(){
+			var city = $scope.cities.pop();
+			$scope.map = { center: { latitude: city.latitude, longitude:  city.longitude}, zoom: 8 };
+			$scope.cities.push(city);
+
+		}
 	}
 ]);
